requirements.txt
# setup_warnme_req.txt
snowflake-connector-python 
pandas
pytz
twilio
